separated inner_tf_coils classes
